Extract addtional ComponentEvidence types
This moves all new evidence related types into their own spec file.
