+++ cyclonedx-bom/src/specs/common/component.rs
@@ -25,14 +25,14 @@ pub(crate) mod base {
 
     #[versioned("1.4")]
     use crate::specs::v1_4::external_reference::ExternalReferences;
+    #[versioned("1.5")]
+    use crate::specs::v1_5::{
+        evidence::{Callstack, Identity, Occurrences},
+        external_reference::ExternalReferences,
+        modelcard::ModelCard,
+    };
     #[versioned("1.3")]
     use crate::{models::bom::SpecVersion, specs::v1_3::external_reference::ExternalReferences};
-    #[versioned("1.5")]
-    use crate::{
-        specs::v1_5::{external_reference::ExternalReferences, modelcard::ModelCard, tool::Tools},
-        utilities::convert_optional_vec,
-        xml::{write_close_tag, write_start_tag},
-    };
 
     use crate::{
         errors::{BomError, XmlReadError},
@@ -917,7 +917,6 @@ pub(crate) mod base {
     }
 
     const EVIDENCE_TAG: &str = "evidence";
-    const CALLSTACK_TAG: &str = "callstack";
 
     impl ToXml for ComponentEvidence {
         fn write_xml_element<W: std::io::Write>(
@@ -1030,390 +1029,6 @@ pub(crate) mod base {
                 #[versioned("1.5")]
                 identity,
             })
-        }
-    }
-
-    #[versioned("1.5")]
-    #[derive(Debug, Deserialize, Serialize, PartialEq)]
-    #[serde(rename_all = "camelCase")]
-    pub(crate) struct Occurrences(Vec<Occurrence>);
-
-    #[versioned("1.5")]
-    const OCCURRENCE_TAG: &str = "occurrence";
-    #[versioned("1.5")]
-    const OCCURRENCES_TAG: &str = "occurrences";
-
-    #[versioned("1.5")]
-    impl ToXml for Occurrences {
-        fn write_xml_element<W: std::io::Write>(
-            &self,
-            writer: &mut xml::EventWriter<W>,
-        ) -> Result<(), crate::errors::XmlWriteError> {
-            self.write_xml_named_element(writer, OCCURRENCES_TAG)
-        }
-    }
-
-    #[versioned("1.5")]
-    impl ToInnerXml for Occurrences {
-        fn write_xml_named_element<W: std::io::Write>(
-            &self,
-            writer: &mut xml::EventWriter<W>,
-            tag: &str,
-        ) -> Result<(), crate::errors::XmlWriteError> {
-            write_start_tag(writer, tag)?;
-
-            for occurrence in &self.0 {
-                occurrence.write_xml_element(writer)?;
-            }
-
-            write_close_tag(writer, tag)?;
-
-            Ok(())
-        }
-    }
-
-    #[versioned("1.5")]
-    impl FromXml for Occurrences {
-        fn read_xml_element<R: std::io::Read>(
-            event_reader: &mut xml::EventReader<R>,
-            element_name: &xml::name::OwnedName,
-            _attributes: &[xml::attribute::OwnedAttribute],
-        ) -> Result<Self, crate::errors::XmlReadError>
-        where
-            Self: Sized,
-        {
-            read_list_tag(event_reader, element_name, OCCURRENCE_TAG).map(Occurrences)
-        }
-    }
-
-    #[versioned("1.5")]
-    impl From<Occurrences> for models::component::Occurrences {
-        fn from(other: Occurrences) -> Self {
-            Self(convert_vec(other.0))
-        }
-    }
-
-    #[versioned("1.5")]
-    impl From<models::component::Occurrences> for Occurrences {
-        fn from(other: models::component::Occurrences) -> Self {
-            Self(convert_vec(other.0))
-        }
-    }
-
-    #[versioned("1.5")]
-    #[derive(Debug, Deserialize, Serialize, PartialEq)]
-    #[serde(rename_all = "camelCase")]
-    pub(crate) struct Occurrence {
-        #[serde(skip_serializing_if = "Option::is_none")]
-        pub bom_ref: Option<String>,
-        pub location: String,
-    }
-
-    #[versioned("1.5")]
-    impl From<Occurrence> for models::component::Occurrence {
-        fn from(other: Occurrence) -> Self {
-            Self {
-                bom_ref: other.bom_ref.map(crate::models::bom::BomReference::new),
-                location: other.location,
-            }
-        }
-    }
-
-    #[versioned("1.5")]
-    impl From<models::component::Occurrence> for Occurrence {
-        fn from(other: models::component::Occurrence) -> Self {
-            Self {
-                bom_ref: other.bom_ref.map(|s| s.0),
-                location: other.location,
-            }
-        }
-    }
-
-    const LOCATION_TAG: &str = "location";
-
-    #[versioned("1.5")]
-    impl ToXml for Occurrence {
-        fn write_xml_element<W: std::io::Write>(
-            &self,
-            writer: &mut xml::EventWriter<W>,
-        ) -> Result<(), crate::errors::XmlWriteError> {
-            let mut start_tag = XmlEvent::start_element(OCCURRENCE_TAG);
-
-            if let Some(bom_ref) = &self.bom_ref {
-                start_tag = start_tag.attr(BOM_REF_ATTR, bom_ref);
-            }
-
-            writer
-                .write(start_tag)
-                .map_err(to_xml_write_error(OCCURRENCE_TAG))?;
-
-            write_simple_tag(writer, LOCATION_TAG, &self.location)?;
-
-            write_close_tag(writer, OCCURRENCE_TAG)?;
-
-            Ok(())
-        }
-    }
-
-    #[versioned("1.5")]
-    impl FromXml for Occurrence {
-        fn read_xml_element<R: std::io::Read>(
-            event_reader: &mut xml::EventReader<R>,
-            element_name: &xml::name::OwnedName,
-            attributes: &[xml::attribute::OwnedAttribute],
-        ) -> Result<Self, crate::errors::XmlReadError>
-        where
-            Self: Sized,
-        {
-            let bom_ref = optional_attribute(attributes, BOM_REF_ATTR);
-            let mut location: Option<String> = None;
-
-            let mut got_end_tag = false;
-            while !got_end_tag {
-                let next_element = event_reader
-                    .next()
-                    .map_err(to_xml_read_error(EVIDENCE_TAG))?;
-                match next_element {
-                    reader::XmlEvent::StartElement { name, .. }
-                        if name.local_name == LOCATION_TAG =>
-                    {
-                        location = Some(read_simple_tag(event_reader, &name)?);
-                    }
-                    reader::XmlEvent::EndElement { name } if &name == element_name => {
-                        got_end_tag = true;
-                    }
-                    unexpected => return Err(unexpected_element_error(element_name, unexpected)),
-                }
-            }
-
-            let location = location
-                .ok_or_else(|| XmlReadError::required_data_missing(LOCATION_TAG, &element_name))?;
-
-            Ok(Self { bom_ref, location })
-        }
-    }
-
-    #[versioned("1.5")]
-    #[derive(Debug, Deserialize, Serialize, PartialEq)]
-    #[serde(rename_all = "camelCase")]
-    pub(crate) struct Callstack(pub Frames);
-
-    #[versioned("1.5")]
-    impl From<Callstack> for models::component::Callstack {
-        fn from(other: Callstack) -> Self {
-            Self(other.0.into())
-        }
-    }
-
-    #[versioned("1.5")]
-    impl From<models::component::Callstack> for Callstack {
-        fn from(other: models::component::Callstack) -> Self {
-            Self(other.0.into())
-        }
-    }
-
-    #[versioned("1.5")]
-    const FRAMES_TAG: &str = "frames";
-    #[versioned("1.5")]
-    const FRAME_TAG: &str = "frame";
-
-    #[versioned("1.5")]
-    impl FromXml for Callstack {
-        fn read_xml_element<R: std::io::Read>(
-            event_reader: &mut xml::EventReader<R>,
-            element_name: &xml::name::OwnedName,
-            _attributes: &[xml::attribute::OwnedAttribute],
-        ) -> Result<Self, crate::errors::XmlReadError>
-        where
-            Self: Sized,
-        {
-            let frames =
-                read_lax_validation_list_tag(event_reader, element_name, FRAMES_TAG).map(Frames)?;
-            Ok(Self(frames))
-        }
-    }
-
-    #[versioned("1.5")]
-    #[derive(Debug, Deserialize, Serialize, PartialEq)]
-    #[serde(rename_all = "camelCase")]
-    pub(crate) struct Frames(pub Vec<Frame>);
-
-    #[versioned("1.5")]
-    impl From<Frames> for models::component::Frames {
-        fn from(other: Frames) -> Self {
-            Self(convert_vec(other.0))
-        }
-    }
-
-    #[versioned("1.5")]
-    impl From<models::component::Frames> for Frames {
-        fn from(other: models::component::Frames) -> Self {
-            Self(convert_vec(other.0))
-        }
-    }
-
-    #[versioned("1.5")]
-    #[derive(Debug, Deserialize, Serialize, PartialEq)]
-    #[serde(rename_all = "camelCase")]
-    pub(crate) struct Frame {
-        #[serde(skip_serializing_if = "Option::is_none")]
-        pub(crate) package: Option<String>,
-        pub(crate) module: String,
-        #[serde(skip_serializing_if = "Option::is_none")]
-        pub(crate) function: Option<String>,
-        #[serde(skip_serializing_if = "Option::is_none")]
-        pub(crate) parameters: Option<Vec<String>>,
-        #[serde(skip_serializing_if = "Option::is_none")]
-        pub(crate) line: Option<u32>,
-        #[serde(skip_serializing_if = "Option::is_none")]
-        pub(crate) column: Option<u32>,
-        #[serde(skip_serializing_if = "Option::is_none")]
-        pub(crate) full_filename: Option<String>,
-    }
-
-    #[versioned("1.5")]
-    impl From<Frame> for models::component::Frame {
-        fn from(other: Frame) -> Self {
-            Self {
-                package: convert_optional(other.package),
-                module: other.module,
-                function: convert_optional(other.function),
-                parameters: convert_optional_vec(other.parameters),
-                line: convert_optional(other.line),
-                column: convert_optional(other.column),
-                full_filename: convert_optional(other.full_filename),
-            }
-        }
-    }
-
-    #[versioned("1.5")]
-    impl From<models::component::Frame> for Frame {
-        fn from(other: models::component::Frame) -> Self {
-            Self {
-                package: convert_optional(other.package),
-                module: other.module,
-                function: convert_optional(other.function),
-                parameters: convert_optional_vec(other.parameters),
-                line: convert_optional(other.line),
-                column: convert_optional(other.column),
-                full_filename: convert_optional(other.full_filename),
-            }
-        }
-    }
-
-    #[versioned("1.5")]
-    impl FromXml for Frame {
-        fn read_xml_element<R: std::io::Read>(
-            event_reader: &mut xml::EventReader<R>,
-            element_name: &xml::name::OwnedName,
-            _attributes: &[xml::attribute::OwnedAttribute],
-        ) -> Result<Self, XmlReadError>
-        where
-            Self: Sized,
-        {
-            let mut got_end_tag = false;
-            while !got_end_tag {
-                let next_element = event_reader.next().map_err(to_xml_read_error(FRAME_TAG))?;
-
-                match next_element {
-                    reader::XmlEvent::EndElement { name } if &name == element_name => {
-                        got_end_tag = true;
-                    }
-                    _ => (),
-                }
-            }
-
-            todo!("")
-        }
-    }
-
-    #[versioned("1.5")]
-    #[derive(Debug, Deserialize, Serialize, PartialEq)]
-    #[serde(rename_all = "camelCase")]
-    pub(crate) struct Identity {
-        pub(crate) field: String,
-        #[serde(skip_serializing_if = "Option::is_none")]
-        pub(crate) confidence: Option<f32>,
-        #[serde(skip_serializing_if = "Option::is_none")]
-        pub(crate) methods: Option<Methods>,
-        #[serde(skip_serializing_if = "Option::is_none")]
-        pub(crate) tools: Option<Tools>,
-    }
-
-    #[versioned("1.5")]
-    impl From<Identity> for models::component::Identity {
-        fn from(other: Identity) -> Self {
-            Self {
-                field: other.field,
-                confidence: other
-                    .confidence
-                    .map(models::component::ConfidenceScore::new),
-                methods: convert_optional(other.methods),
-                tools: convert_optional(other.tools),
-            }
-        }
-    }
-
-    #[versioned("1.5")]
-    impl From<models::component::Identity> for Identity {
-        fn from(other: models::component::Identity) -> Self {
-            Self {
-                field: other.field,
-                confidence: other.confidence.map(|s| s.get()),
-                methods: convert_optional(other.methods),
-                tools: try_convert_optional(other.tools).expect("Failed to convert"),
-            }
-        }
-    }
-
-    #[versioned("1.5")]
-    #[derive(Debug, Deserialize, Serialize, PartialEq)]
-    #[serde(rename_all = "camelCase")]
-    pub(crate) struct Methods(Vec<Method>);
-
-    #[versioned("1.5")]
-    impl From<Methods> for models::component::Methods {
-        fn from(other: Methods) -> Self {
-            Self(convert_vec(other.0))
-        }
-    }
-
-    #[versioned("1.5")]
-    impl From<models::component::Methods> for Methods {
-        fn from(other: models::component::Methods) -> Self {
-            Self(convert_vec(other.0))
-        }
-    }
-
-    #[versioned("1.5")]
-    #[derive(Debug, Deserialize, Serialize, PartialEq)]
-    #[serde(rename_all = "camelCase")]
-    pub(crate) struct Method {
-        pub(crate) technique: String,
-        pub(crate) confidence: f32,
-        #[serde(skip_serializing_if = "Option::is_none")]
-        pub(crate) value: Option<String>,
-    }
-
-    #[versioned("1.5")]
-    impl From<Method> for models::component::Method {
-        fn from(other: Method) -> Self {
-            Self {
-                technique: other.technique,
-                confidence: models::component::ConfidenceScore::new(other.confidence),
-                value: convert_optional(other.value),
-            }
-        }
-    }
-
-    #[versioned("1.5")]
-    impl From<models::component::Method> for Method {
-        fn from(other: models::component::Method) -> Self {
-            Self {
-                technique: other.technique,
-                confidence: other.confidence.get(),
-                value: convert_optional(other.value),
-            }
         }
     }
 
@@ -1731,6 +1346,10 @@ pub(crate) mod base {
         };
         #[versioned("1.5")]
         use crate::specs::v1_5::{
+            evidence::test::{
+                corresponding_callstack, corresponding_identity, corresponding_occurrences,
+                example_callstack, example_identity, example_occurrences,
+            },
             external_reference::test::{
                 corresponding_external_references, example_external_references,
             },
@@ -1952,80 +1571,6 @@ pub(crate) mod base {
             models::component::CopyrightTexts(vec![corresponding_copyright()])
         }
 
-        #[versioned("1.5")]
-        fn example_occurrences() -> Occurrences {
-            Occurrences(vec![Occurrence {
-                bom_ref: Some("occurrence-1".to_string()),
-                location: "location-1".to_string(),
-            }])
-        }
-
-        #[versioned("1.5")]
-        fn corresponding_occurrences() -> models::component::Occurrences {
-            models::component::Occurrences(vec![models::component::Occurrence {
-                bom_ref: Some(models::bom::BomReference::new("occurrence-1")),
-                location: "location-1".to_string(),
-            }])
-        }
-
-        #[versioned("1.5")]
-        fn example_callstack() -> Callstack {
-            Callstack(Frames(vec![Frame {
-                package: Some("package-1".to_string()),
-                module: "module-1".to_string(),
-                function: Some("function".to_string()),
-                parameters: None,
-                line: Some(10),
-                column: Some(20),
-                full_filename: Some("full-filename".to_string()),
-            }]))
-        }
-
-        #[versioned("1.5")]
-        fn corresponding_callstack() -> models::component::Callstack {
-            models::component::Callstack(models::component::Frames(vec![
-                models::component::Frame {
-                    package: Some("package-1".to_string()),
-                    module: "module-1".to_string(),
-                    function: Some("function".to_string()),
-                    parameters: None,
-                    line: Some(10),
-                    column: Some(20),
-                    full_filename: Some("full-filename".to_string()),
-                },
-            ]))
-        }
-
-        #[versioned("1.5")]
-        fn example_identity() -> Identity {
-            Identity {
-                field: "field".to_string(),
-                confidence: Some(0.5),
-                methods: Some(Methods(vec![Method {
-                    technique: "technique-1".to_string(),
-                    confidence: 0.8,
-                    value: Some("identity-value".to_string()),
-                }])),
-                tools: None,
-            }
-        }
-
-        #[versioned("1.5")]
-        fn corresponding_identity() -> models::component::Identity {
-            models::component::Identity {
-                field: "field".to_string(),
-                confidence: Some(models::component::ConfidenceScore::new(0.5)),
-                methods: Some(models::component::Methods(vec![
-                    models::component::Method {
-                        technique: "technique-1".to_string(),
-                        confidence: models::component::ConfidenceScore::new(0.8),
-                        value: Some("identity-value".to_string()),
-                    },
-                ])),
-                tools: None,
-            }
-        }
-
         fn example_copyright() -> Copyright {
             Copyright {
                 text: "copyright".to_string(),
@@ -2034,98 +1579,6 @@ pub(crate) mod base {
 
         fn corresponding_copyright() -> models::component::Copyright {
             models::component::Copyright("copyright".to_string())
-        }
-
-        #[versioned("1.5")]
-        #[test]
-        fn it_should_read_xml_occurrences() {
-            let input = r#"
-<occurrences>
-  <occurrence bom-ref="d6bf237e-4e11-4713-9f62-56d18d5e2079">
-    <location>/path/to/component</location>
-  </occurrence>
-  <occurrence bom-ref="b574d5d1-e3cf-4dcd-9ba5-f3507eb1b175">
-    <location>/another/path/to/component</location>
-  </occurrence>
-</occurrences>
-"#;
-            let actual: Occurrences = read_element_from_string(input);
-            let expected = Occurrences(vec![
-                Occurrence {
-                    bom_ref: Some("d6bf237e-4e11-4713-9f62-56d18d5e2079".to_string()),
-                    location: "/path/to/component".to_string(),
-                },
-                Occurrence {
-                    bom_ref: Some("b574d5d1-e3cf-4dcd-9ba5-f3507eb1b175".to_string()),
-                    location: "/another/path/to/component".to_string(),
-                },
-            ]);
-            assert_eq!(actual, expected);
-        }
-
-        #[versioned("1.5")]
-        #[test]
-        fn it_should_write_xml_occurrences() {
-            let xml_output = write_element_to_string(example_occurrences());
-            insta::assert_snapshot!(xml_output);
-        }
-
-        #[versioned("1.5")]
-        #[test]
-        fn it_should_read_xml_callstack() {
-            let input = r#"
-<callstack>
-  <frames>
-    <frame>
-      <package>com.apache.logging.log4j.core</package>
-      <module>Logger.class</module>
-      <function>logMessage</function>
-      <parameters>
-        <parameter>com.acme.HelloWorld</parameter>
-        <parameter>Level.INFO</parameter>
-        <parameter>null</parameter>
-        <parameter>Hello World</parameter>
-      </parameters>
-      <line>150</line>
-      <column>17</column>
-      <fullFilename>full-filename.java</fullFilename>
-    </frame>
-    <frame>
-      <module>HelloWorld.class</module>
-      <function>main</function>
-      <line>20</line>
-      <column>12</column>
-      <fullFilename>/path/to/HelloWorld.class</fullFilename>
-    </frame>
-  </frames>
-</callstack>"#;
-            let actual: Callstack = read_element_from_string(input);
-            let expected = Callstack(Frames(vec![
-                Frame {
-                    package: Some("com.apache.logging.log4j.core".to_string()),
-                    module: "Logger.class".to_string(),
-                    function: Some("logMessage".to_string()),
-                    parameters: Some(vec![
-                        "com.acme.HelloWorld".to_string(),
-                        "Level.INFO".to_string(),
-                        "null".to_string(),
-                        "Hello World".to_string(),
-                    ]),
-                    line: Some(150),
-                    column: Some(17),
-                    full_filename: Some("full-filename.java".to_string()),
-                },
-                Frame {
-                    package: None,
-                    module: "HelloWorld.class".to_string(),
-                    function: None,
-                    parameters: None,
-                    line: Some(20),
-                    column: Some(12),
-                    full_filename: Some("/path/to/HelloWorld.class".to_string()),
-                },
-            ]));
-            assert_eq!(actual, expected);
         }
 
         #[test]
